Extract `{update,delete}_all_test.rb` from `persistence_test.rb` and `relations_test.rb`
`persistence_test.rb` and `relations_test.rb` have too many lines, so
I'd like to extract relation around tests to dedicated files before
newly test added.
